Music Selection App
This is a Java program that uses python scripts, a million song data set
and H5 files to extract songs and play them for the user. A series of
questions are asked to the user to find out the users likes and dislikes
and recommends another song to help the user find a song that they like.
We have used categories like tempo, artist and loudness to help choose a
better song choice.